third_party/WebKit/LayoutTests/bluetooth/requestDevice/acceptAllDevices/device-with-no-name.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Device with no name or UUIDs nearby. Should be found if ' +
   'acceptAllDevices is true.';

bluetooth_test(() => getNamelessHealthThermometerDevice()
  .then(() => requestDeviceWithTrustedClick({acceptAllDevices: true}))
  .then(device => assert_true(device.name === null)),
  test_desc);
</script>
